Invalidate spurious failure of CAS_PTR
Invalidate the spurious failure of CAS_PTR  by setting the weak
argument of __atomic_compare_exchange to 0.

In original implementation of CAS_PTR, it allows spurious failure,
which means it fails even if the comparison is equal. The caller can
not determine whether the CAS_PTR success or not because it return the
value only by the result of comparison. In include/lock.h , the
function lock_lock use below code to implement lock:

  while (CAS_U32(l, (uint32_t) 0, (uint32_t) 1) == 1)
      ;

In case the CAS_PTR fails even if *l == 0, the function lock_lock could
spuriously consider it acquires the lock successfully because CAS_PTR
returns 0, but actually, it doesn't acquire the lock.